requirements.txt
flask
flask_cors
flask-WTF
pandas
flask-sqlalchemy
python-dotenv
email_validator